Crate nargo_doc

Modules

html 🔒

ids 🔒

items

Contains the data structures that represent items to be shown by the documentation generator.

links

Structs

BrokenLink

ConvertedItem

DocItemBuilder 🔒

Functions

convert_primitive_type 🔒

crate_module

Returns the root module in a crate.

link_offset 🔒

Returns the offset in text at which the actual comment text start at the given line, assuming text is the entire doc comment text. The start and end position of links are relative to the line in which a link appears, and because broken links need to be reported as [Location]s we need to adjust their span accordingly.

to_html

Returns a list of (path, contents) representing the HTML files for the given crates. The paths are relative paths that can be joined to a base directory.