nargo_doc::html::has_uri
Trait HasUri

pub(super) trait HasUri {
    // Required method
    fn uri(&self) -> String;
}

A trait that associates an HTML uri with an item. The uri is relative to the page where the item is typically shown (for example, a crate is shown in the workspace, a struct is shown in its module, etc.).

Required Methods

fn uri(&self) -> String