nargo_doc::html
Struct HTMLCreator

struct HTMLCreator {
    output: String,
    files: Vec<(PathBuf, String)>,
    current_path: Vec<String>,
    current_crate_version: Option<String>,
    workspace_name: String,
    id_to_info: HashMap<ItemId, ItemInfo>,
    all_trait_impls: HashMap<ItemId, HashSet<TraitImpl>>,
    self_type: Option<Type>,
}

Builds all the files for the HTML documentation.

Fields
output: Stringfiles: Vec<(PathBuf, String)>current_path: Vec<String>current_crate_version: Option<String>workspace_name: Stringid_to_info: HashMap<ItemId, ItemInfo>all_trait_impls: HashMap<ItemId, HashSet<TraitImpl>>

Maps a trait ID to all its implementations across all crates.

self_type: Option<Type>
Implementations

impl HTMLCreator

fn new(workspace: &Workspace) -> Self

fn process_workspace(&mut self, workspace: &Workspace)

fn create_styles(&mut self)

fn create_js(&mut self)

fn create_all_items(&mut self, workspace: &Workspace)

fn render_all_items_list( &mut self, title: &str, class: &str, items: &[(Vec<String>, String)], )

fn render_all_items_sidebar(&mut self, all_items: &AllItems)

fn create_index(&mut self, workspace: &Workspace)

fn create_crate(&mut self, workspace: &Workspace, krate: &Crate)

fn render_crate_sidebar(&mut self, workspace: &Workspace, krate: &Crate)

fn render_items( &mut self, items: &[(ItemVisibility, Item)], sidebar: bool, nesting: usize, )

fn render_reexports(&mut self, items: &[(ItemVisibility, Item)])

fn render_contracts( &mut self, items: &[(ItemVisibility, Item)], sidebar: bool, nesting: usize, )

fn render_modules( &mut self, items: &[(ItemVisibility, Item)], sidebar: bool, nesting: usize, )

fn render_structs( &mut self, items: &[(ItemVisibility, Item)], sidebar: bool, nesting: usize, )

fn render_traits( &mut self, items: &[(ItemVisibility, Item)], sidebar: bool, nesting: usize, )

fn render_type_aliases( &mut self, items: &[(ItemVisibility, Item)], sidebar: bool, nesting: usize, )

fn render_primitive_types( &mut self, items: &[(ItemVisibility, Item)], sidebar: bool, nesting: usize, )

fn render_globals( &mut self, items: &[(ItemVisibility, Item)], sidebar: bool, nesting: usize, )

fn render_functions( &mut self, items: &[(ItemVisibility, Item)], sidebar: bool, nesting: usize, )

fn render_list<T: ItemProperties + HasUri + HasClass>( &mut self, title: &str, anchor: &str, sidebar: bool, nesting: usize, items: &[&T], )

fn render_reexports_list( &mut self, title: &str, anchor: &str, reexports: &[&Reexport], )

fn create_items( &mut self, parent_module: &Module, items: &[(ItemVisibility, Item)], )

fn create_item(&mut self, parent_module: &Module, item: &Item)

fn create_module(&mut self, parent_module: &Module, module: &Module)

fn render_module_sidebar(&mut self, parent_module: &Module, module: &Module)

fn render_module_items_sidebar(&mut self, title: &str, module: &Module)

fn render_module_contents_sidebar(&mut self, module: &Module, nesting: usize)

fn create_struct(&mut self, parent_module: &Module, struct_: &Struct)

fn render_struct_sidebar(&mut self, struct_: &Struct)

fn create_trait(&mut self, parent_module: &Module, trait_: &Trait)

fn render_trait_sidebar(&mut self, trait_: &Trait)

fn render_trait_sidebar_methods(&mut self, title: &str, methods: &[Function])

fn render_sidebar_trait_impls( &mut self, title: &str, display_trait: bool, trait_impls: &[TraitImpl], )

fn create_alias(&mut self, parent_module: &Module, alias: &TypeAlias)

fn create_function(&mut self, parent_module: &Module, function: &Function)

fn create_global(&mut self, parent_module: &Module, global: &Global)

fn create_primitive_type( &mut self, parent_module: &Module, primitive: &PrimitiveType, )

fn render_primitive_sidebar(&mut self, primitive: &PrimitiveType)

fn render_struct_code(&mut self, struct_: &Struct)

fn render_struct_fields(&mut self, fields: &[StructField])

fn render_impls(&mut self, impls: &[Impl])

fn render_impl(&mut self, impl_: &Impl)

fn render_trait_impls( &mut self, trait_impls: &[TraitImpl], title: &str, show_methods: bool, )

fn render_trait_impl(&mut self, trait_impl: &TraitImpl, show_methods: bool)

fn render_trait_code(&mut self, trait_: &Trait)

fn render_trait_methods(&mut self, title: &str, methods: &[Function])

fn render_type_alias_code(&mut self, alias: &TypeAlias)

fn render_global_code(&mut self, global: &Global)

fn render_methods( &mut self, methods: &[Function], current_heading_level: usize, output_id: bool, link: bool, )

fn render_function( &mut self, function: &Function, current_heading_level: usize, as_header: bool, link: bool, output_id: bool, )

fn render_function_signature( &mut self, function: &Function, as_header: bool, link: bool, output_id: bool, )

fn render_function_signature_inner( &mut self, function: &Function, color_name: bool, link: bool, indent: usize, )

fn render_generics(&mut self, generics: &[Generic])

fn render_where_clause( &mut self, where_clause: &[TraitConstraint], indent: usize, )

fn render_trait_bound(&mut self, bound: &TraitBound)

fn render_id_reference(&mut self, id: &ItemId, name: &str)

fn render_type(&mut self, typ: &Type)

fn type_to_string(&self, typ: &Type) -> String

fn render_generic_types(&mut self, generics: &[Type])

fn render_trait_generics( &mut self, ordered: &[Type], named: &BTreeMap<String, Type>, )

fn render_comments( &mut self, comments: Option<&Comments>, current_heading_level: usize, )

Renders the given comments, if any. Markdown headers that are less or equal than current_heading_level will be rendered as smaller headers (i.e., ### becomes #### if current_heading_level is 3) to ensure proper nesting.

fn process_comments_links(&self, links: &Links, comments: String) -> String

Replace links in the given comments with proper HTML links to pages and anchors related to target items.

fn render_breadcrumbs(&mut self, last_is_link: bool)

fn get_all_trait_impls(&self, trait_: &Trait) -> Vec<TraitImpl>

fn html_start(&mut self, title: &str, short_title: &str)

fn html_start_with_redirect( &mut self, title: &str, short_title: &str, redirect: Option<String>, )

fn html_end(&mut self)

fn main_start(&mut self, last_breadcrumb_is_link: bool)

fn main_end(&mut self)

fn sidebar_start(&mut self)

fn sidebar_end(&mut self)

fn h1(&mut self, text: &str)

fn h2(&mut self, text: &str)

fn h3(&mut self, text: &str)

fn push_file(&mut self, path: PathBuf)