nargo_doc

Module items

Contains the data structures that represent items to be shown by the documentation generator.

Structs

AssociatedConstant

AssociatedType

Crate

Function

FunctionParam

Generic

Global

Impl

ItemId

Uniquely identifies an item. This is done by using a type’s name, location in source code and kind. With macros, two types might end up being defined in the same location but they will likely have different names. This is just a temporary solution until we have a better way to uniquely identify items across crates.

Link

A markdown link that resolves to an item, in one of these forms:

Module

PrimitiveType

Reexport

Struct

StructField

Trait

TraitBound

TraitConstraint

TraitImpl

TypeAlias

Workspace

Enums

Item

ItemKind

LinkTarget

PrimitiveTypeKind

Type

Traits

ItemProperties

Type Aliases

Comments

Links