nargo_doc
Module links

Structs

Link

A resolved markdown link found in a line of markdown, in the form:

LinkFinder

Finds links in markdown comments.

PlainLink 🔒

An unresolved markdown link found in a line of markdown, in the form:

Enums

CurrentType

LinkTarget

Functions

find_links_in_markdown_line 🔒

path_to_link_target 🔒

Tries to convert a path into a link by resolving a path like std::collections::Vec. This is similar to how name resolution works in the compiler, except that it’s simpler (no need to report errors), and references to type and trait functions are handled a bit differently.

path_to_link_target_searching_modules 🔒

primitive_type_method_link_target 🔒

primitive_type_or_primitive_type_method_link_target 🔒

reference_regex 🔒

A regex that captures markdown links as either [reference], [reference][link] or [reference](url), ignoring those that happen inside backticks.

trait_method_link_target 🔒

type_method_or_field_link_target 🔒